requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.20.37
botocore==1.23.37
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
install==1.3.5
jmespath==0.10.0
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
